refactor: update client
